Change UpdateRole to Update to match naming conventions accross application
